pyproject.toml
[project]
name = "jove-nvim"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "ipywidgets>=8.1.7",
    "jupyter-client>=8.6.3",
    "matplotlib>=3.10.6",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "rich>=14.1.0",
]
